backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.4.0
httpx==0.27.2
openai-whisper==20231117
gTTS==2.5.1
APScheduler==3.10.4
twilio==9.4.5
